test: update design electric record tests
